+++ gui_engine/server/gui_api.c
@@ -283,8 +283,9 @@ void *gui_realloc(void *ptr_old, uint32_t n)
 
 void gui_free(void *rmem)
 {
-    if (((uint32_t)rmem >= (uint32_t)os_api->lower_mem_addr) && \
-        ((uint32_t)rmem <= (uint32_t)os_api->lower_mem_addr + (uint32_t)os_api->lower_mem_size)
+    if (((uint32_t)(uintptr_t)rmem >= (uint32_t)(uintptr_t)os_api->lower_mem_addr) && \
+        ((uint32_t)(uintptr_t)rmem <= (uint32_t)(uintptr_t)os_api->lower_mem_addr +
+         (uint32_t)os_api->lower_mem_size)
        )
     {
         gui_lower_free(rmem);
